feat: seo child pages
